requirements.txt
asgiref==3.5.2
Django==4.0.6
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
